Userstory/rebel action card data setup

+++ Component-Info-API/src/main/java/com/starwars/rebellion/ComponentInfoAPI/repositorys/LeaderRepository.java
@@ -5,7 +5,7 @@ import com.starwars.rebellion.ComponentInfoAPI.dao.Leader;
 import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
-public interface LeaderRepository extends CrudRepository<Leader, Long> {
+public interface LeaderRepository extends CrudRepository<Leader, Integer> {
     Leader findByName(String name);
     List<Leader> findAllByFaction(Faction faction);
 }
